Try to move syncing stuff from `sc-network-common` to `sc-network-sync`
