+++ substrate/client/informant/Cargo.toml
@@ -19,6 +19,7 @@ futures-timer = "3.0.1"
 log = "0.4.17"
 sc-client-api = { path = "../api" }
 sc-network-common = { path = "../network/common" }
+sc-network-sync = { path = "../network/sync" }
 sc-network = { path = "../network" }
 sp-blockchain = { path = "../../primitives/blockchain" }
 sp-runtime = { path = "../../primitives/runtime" }
