+++ substrate/client/network-gossip/src/bridge.rs
@@ -22,7 +22,7 @@ use crate::{
 };
 
 use sc_network::{event::Event, types::ProtocolName, ReputationChange};
-use sc_network_common::sync::SyncEvent;
+use sc_network_sync::SyncEvent;
 
 use futures::{
 	channel::mpsc::{channel, Receiver, Sender},
@@ -338,7 +338,8 @@ mod tests {
 		config::MultiaddrWithPeerId, NetworkBlock, NetworkEventStream, NetworkNotification,
 		NetworkPeers, NotificationSenderError, NotificationSenderT as NotificationSender,
 	};
-	use sc_network_common::{role::ObservedRole, sync::SyncEventStream};
+	use sc_network_common::role::ObservedRole;
+	use sc_network_sync::SyncEventStream;
 	use sp_runtime::{
 		testing::H256,
 		traits::{Block as BlockT, NumberFor},
