+++ substrate/client/network-gossip/src/lib.rs
@@ -71,7 +71,7 @@ use libp2p::{multiaddr, PeerId};
 use sc_network::{
 	types::ProtocolName, NetworkBlock, NetworkEventStream, NetworkNotification, NetworkPeers,
 };
-use sc_network_common::sync::SyncEventStream;
+use sc_network_sync::SyncEventStream;
 use sp_runtime::traits::{Block as BlockT, NumberFor};
 use std::iter;
