+++ substrate/client/network/statement/Cargo.toml
@@ -21,6 +21,7 @@ libp2p = "0.51.3"
 log = "0.4.17"
 prometheus-endpoint = { package = "substrate-prometheus-endpoint", path = "../../../utils/prometheus" }
 sc-network-common = { path = "../common" }
+sc-network-sync = { path = "../sync" }
 sc-network = { path = ".." }
 sp-consensus = { path = "../../../primitives/consensus/common" }
 sp-statement-store = { path = "../../../primitives/statement-store" }
