+++ substrate/client/service/src/metrics.rs
@@ -23,7 +23,7 @@ use futures_timer::Delay;
 use prometheus_endpoint::{register, Gauge, GaugeVec, Opts, PrometheusError, Registry, U64};
 use sc_client_api::{ClientInfo, UsageProvider};
 use sc_network::{config::Role, NetworkStatus, NetworkStatusProvider};
-use sc_network_common::sync::{SyncStatus, SyncStatusProvider};
+use sc_network_sync::{SyncStatus, SyncStatusProvider};
 use sc_telemetry::{telemetry, TelemetryHandle, SUBSTRATE_INFO};
 use sc_transaction_pool_api::{MaintainedTransactionPool, PoolStatus};
 use sc_utils::metrics::register_globals;
